Move more tests to source tree
