+++ src/diffTool/rulesProcessor.test.ts
